backend/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
joblib
python-docx
gunicorn